add: 3D live animation of tracked targets
The tracker application now renders reflection points on a 3D plot which is animated in real time. It also plots the designation of tracking area, box frames around detected targets and movement trails of tracked targets.

This has been tested with the evaluation module IWR6843AOP EVM Rev. F (ES2.0), which is running the out of the box demo (SDK v3.5.0.4). Binaries are in a separate folder within this repository.

The configuration file used to configure the radar is best_velocity_res-no_grouping.cfg